Fixed carrier frequency problems.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@2699 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/openpilotgcs/src/plugins/pipxtreme/pipxtremegadgetwidget.cpp
@@ -494,7 +494,7 @@ void PipXtremeGadgetWidget::saveToFlash()
 
 	s = m_widget->doubleSpinBox_Frequency->text().trimmed();
 	s.replace(' ', "");	// remove all spaces
-	settings.frequency_Hz = s.toFloat(&ok) * 1e6;
+	settings.frequency_Hz = (uint32_t)(s.toDouble(&ok) * 1e6);
 	if (s.isEmpty() || !ok || settings.frequency_Hz < settings.min_frequency_Hz || settings.frequency_Hz > settings.max_frequency_Hz)
 	{
 		error("Check your \"Frequency\" entry!", 0);
